+++ src/configui/app/config-options/advanced-videoconfig/advanced-videoconfig.component.ts
@@ -47,6 +47,7 @@ export class AdvancedVideoconfigComponent
   maxHeight: number | undefined = undefined;
   maxFPS: number | undefined = undefined;
   maxBitrate: number | undefined = undefined;
+  useOneProcess: boolean | undefined = undefined;
 
   async readValue() {
     const config = await this.getCameraConfig(this.accessory?.uniqueId || '');
@@ -122,6 +123,9 @@ export class AdvancedVideoconfigComponent
     }
     if (this.maxBitrate !== undefined) {
       newConfig['maxBitrate'] = this.maxBitrate;
+    }
+    if (this.useOneProcess) {
+      newConfig['useOneProcess'] = this.useOneProcess;
     }
 
     this.updateConfig(